Make three instances of "mounted_backup" distinguishable
The identifier "mounted_backup" is used as three different things in the
script: As id of the drive that is attached from PBS, as id of the
devices that is attached to the VM, and as serial number of said device
(for use in VM).

This commit makes that distinction explicit by using variables.